+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # v2.17.3
 
+* changed the behaviour of `Closeable` to adjust
+  for semi-initialized classes